refactored config dir to include a seperate type_settings.py
