+++ db_api/config/__init__.py
@@ -1,0 +1,6 @@
+"""
+Configuration package for the Database API.
+
+This package contains modules related to application configuration, including raw settings
+and type-validated settings.
+"""
