+++ db_api/utils/jwt_utils.py
@@ -9,7 +9,7 @@ import jwt
 from typing import Dict, Any
 from fastapi import HTTPException, Depends
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
-from db_api.config.settings import JWT_SECRET
+from db_api.config.type_settings import JWT_SECRET
 
 def get_current_user(
     credentials: HTTPAuthorizationCredentials = Depends(HTTPBearer())
